+++ app/Controllers/InfoController.js
@@ -8,8 +8,13 @@ const sendInfo = catchAsync(async (_req, res) => {
   res.send({
     identifier: 'vocascan-server',
     version: getVersion(),
-    locked: config.server.registration_locked,
+    locked: config.service.invite_code,
     commitRef: gitDescribe === undefined ? (gitDescribe = await getGitDescribe()) : gitDescribe,
+    email_confirm: {
+      enabled: config.service.email_confirm,
+      level: config.service.email_confirm_level,
+      time: config.service.email_confirm_time,
+    },
   });
 });