+++ app/config/joi.js
@@ -1,3 +1,5 @@
+const ms = require('ms');
+
 const Joi = require('joi')
   .extend((joi) => ({
     base: joi.array(),
@@ -14,6 +16,19 @@ const Joi = require('joi')
         return { value: Object.values(value) };
       }
       return { value: [value] };
+    },
+  }))
+  .extend((joi) => ({
+    base: joi.number(),
+    type: 'ms',
+    coerce: (value) => {
+      return { value: ms(value) };
+    },
+    overrides: {
+      default(...args) {
+        const value = this.$_parent('default', ...args).$_getFlag('default');
+        return this.$_setFlag('default', ms(value));
+      },
     },
   }));